refactor: unify badge and card helpers
