+++ components/ThemeToggle.jsx
@@ -6,7 +6,7 @@ export default function ThemeToggle() {
   if (!mounted) return null;
   return (
     <button onClick={toggle} aria-label="Toggle theme"
-      className="px-3 py-2 rounded-full border hover:bg-white dark:hover:bg-slate-800 flex items-center gap-2">
+      className="pill-outline flex items-center gap-2">
       {dark ? <Sun className="w-4 h-4" /> : <Moon className="w-4 h-4" />}
       <span className="hidden sm:inline">{dark ? "Light" : "Dark"}</span>
     </button>
