+++ styles/globals.css
@@ -14,6 +14,9 @@ body { @apply bg-gradient-to-b from-white to-slate-50 text-slate-900; }
 /* Shared */
 .card { box-shadow: 0 1px 3px rgba(0,0,0,.08), 0 1px 2px rgba(0,0,0,.06); }
 .badge { @apply inline-flex items-center px-3 py-1 rounded-full text-sm bg-slate-100; }
+.pill { @apply px-3 py-2 rounded-full; }
+.pill-outline { @apply pill border hover:bg-white dark:hover:bg-slate-800; }
+.section-container { @apply max-w-5xl mx-auto px-4 sm:px-6 lg:px-8; }
 
 /* Hide scrollbars but allow scrolling */
 .no-scrollbar {
